* rust-sdk: (305 commits)
  Update breez-sdk dependency
  remove packages dir
  work arround grenlight issue with miscaculation of spendable amount
  fix udl file
  Document version of uniffi_bindgen
  add fields to swap_info and some more tests
  Fix validation error message
  Connect LNURL pay to the UI (#291)
  Extract error message from payment error (#283)
  Rename the struct
  Re-generate bridge files
  Rename and document LnUrlPayOperationResult struct
  Update README.md
  update README
  fix types in  readme
  android binding tests
  Fix generated files removing the override bool type (#290)
  Fix string to byte conversion (#282)
  add android build and binding to the make file.
  Rename binding::pay to pay_lnurl
  ...

+++ lib/routes/withdraw_funds/withdraw_funds_address_next_button.dart
@@ -3,12 +3,11 @@ import 'package:c_breez/routes/withdraw_funds/withdraw_funds_confirmation_page.d
 import 'package:c_breez/widgets/route.dart';
 import 'package:c_breez/widgets/single_button_bottom_bar.dart';
 import 'package:flutter/material.dart';
-import 'package:fixnum/fixnum.dart';
 
 class WithdrawFundsAddressNextButton extends StatelessWidget {
   final TextEditingController addressController;
   final bool Function() validator;
-  final Int64 amount;
+  final int amount;
 
   const WithdrawFundsAddressNextButton({
     Key? key,
